Begin rewrite for real-time with signal.asm
This covers basic, low-level observer routines for getting signals
from the IR sensor, in a variety of forms to account for the exact needs
of different use-cases.

These could be heavily optimised for accuracy in exchange for complexity and
total runtime, but for now we're using the simplest versions we can.